Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg[binary]==3.3.4
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
marshmallow==3.20.1
resend==2.0.0
pytest==7.4.3
pytest-env==1.1.3